细节处理

+++ app/src/main/java/com/td/oldplay/bean/UserBean.java
@@ -24,7 +24,7 @@ public class UserBean  implements Serializable{
 
     public String avatar;
     public int classHour;
-    public int money;
+    public float money;
     public String nickName;
     public String password;
     public String payPassword;
